Add results, screenshots api and use mysql

+++ api/routes/results.js
@@ -3,21 +3,38 @@ const router = express.Router();
 const Result = require('../models/result');
 
 router.get('/', (req, res, next) => {
-  Result.findAll()
-  .then((result) => {
-    res.json(result);
+  const jobId = req.query.jobId;
+  let promise;
+  if (jobId) {
+    promise = Result.find({ jobId, deep: true });
+  } else {
+    promise = Result.findAll();
+  }
+  return promise
+  .then((resultOrResults) => {
+    res.json(resultOrResults);
   })
   .catch(next);
 });
 
 router.delete('/:id', (req, res, next) => {
-  const id = req.params.id;
-  Result.find(id)
+  const resultId = req.params.id;
+  Result.find({ resultId, deep: false })
   .then((result) => {
-    return result.delete();
+    if (!result) { return; }
+    return result.delete()
   })
-  .then((r) => {
+  .then(() => {
     res.status(204).end();
+  })
+  .catch(next);
+});
+
+router.post('/', (req, res, next) => {
+  const result = new Result(req.body);
+  result.save()
+  .then(() => {
+    res.status(201).end();
   })
   .catch(next);
 });
